pacdiff-merge3: Remove hardcoded extensions and download feature
With the release of pacman 5.2, packages may now be compressed with zstd
instead of xz. This means that custom scripts (like this one) can no
longer rely on the extension of all package files being `.tar.xz`,
because it could be `.tar.zst` instead.

As a consequence, the file name for a given package can no longer be
fully constructed from package info (name, version, arch). This slightly
complicates the implementation of the script's main functionality, but
it makes downloading non-cached versions from the Arch Linux Archive
especially annoying, because (as far as i can tell) ALA doesn't provide
a machine-friendly way to find a package by name and version, without
knowing the file extension. The only solutions I can think of would be
to:
 - scrape the HTML directory listings;
 - try both `.tar.xz` and `.tar.zst` extensions (but this might break
   again, if more package formats are added in the future).

Since I usually keep at least one previous version of all packages in
cache, I never really need the download feature, therefore I've chosen
to simply remove it, and make the script exit with an error message if
the previous version is not available. In this case, it can still be
downloaded manually, if required.