spacemacs: Fix sql-indent configuration
As of commit syl20bnr/spacemacs@8ded639, spacemacs installs the sql-indent
package from GitHub (at the time of the commit, this was required to override a
different package with the same name, available on MELPA), but, in doing so,
only includes the main sql-indent.el file, leaving out sql-indent-left.el, which
my configuration depends on.

As a fix, copy the relevant parts of the file into my config.