spacemacs: Create interactive update script
Currently, this script only starts a package update after updating
spacemacs itself.

The reason is that updating packages when there is no
emacs daemon running is simple, and, for a spacemacs update, the daemon
must be stopped anyway.
On the other hand, updating packages when the daemon may be running
would require either
 - forcing the user to needlessly stop the daemon
 - using emacsclient to send the update command to the existing daemon
   (which would add complexity to the script)
and manually starting the update is not especially tedious, anyway.

Therefore, I currently think that it's not worth automating package
updates when not also updating spacemacs.