xmobar: Make Wi-Fi status work on laptop after driver switch
Due to an issue in the brcmsmac driver that was causing kernel panics on
my laptop, I've had to switch to the broadcom-wl driver. This driver
doesn't seem to work with the xmobar (nl80211-based) Wireless status
plugin. The simplest solution is stop using this plugin, and make xmobar
call a bespoke wifi-status script instead.