dunst: Remake config from v1.7.3 dunstrc template
I had never updated the dunst config (dunstrc) after initially adding it
to this repo in 2017, so it didn't reflect changes in the config format
since at least dunst v1.3.0. Considering that the config format has
changed a lot since then, and that my config was almost identical to the
default, it was easier to create an up-to-date config by starting from
the current dunstrc template then to update the old config.