.xinitrc: Terminate background processes when logging out
This should ensure that the .xinitrc script doesn't start duplicate instances of
the same background processes when logging out and back in without rebooting.
Most (if not all) of the background processes I start seem to either abort their
startup when they detect that another instance is already running or stop anyway
when they lose their connection to X upon logout, so I didn't actually have any
significant issues, but to me the behavior introduced by this commit feels "more
correct".

As implemented here, termination only works for background processes that don't
fork to background. For this reason, the emacs daemon is now started using the
`--fg-daemon` argument instead of `--daemon`.