ssh: Fix personal/work GitLab key selection
Multiple `IdentityFile` entries don't override eachother: instead, each
command entry adds a key to a list of keys that the SSH client can
attempt to use. Therefore, if the default key is specified globally (for
all hosts), the client may select it instead of the work key, thus
logging in as the wrong GitLab user (because GitLab uses the key to
determine the user, while the actuall SSH username is always `git`).
Specifying `IdentityFile` entries separately for `gitlab.com` and
`gitlab.com-qpq` prevents this.

Another option that's required to solve this problem is `IdentitiesOnly
yes`, which stops the client from using keys which are not specified in
the config but have been added to the SSH agent.

For all hosts other than GitLab, my personal key should automatically be
selected because the identity file has one of the names that are
recognized by default, and all keys that have already been added to the
agent should also be available. Any other hosts which recognize multiple
keys but only accept a specific one will need to be configured like
GitLab.