nvim: Set leader key to space
The previous leader key, 'ò', was easy to press on an Italian keyboard,
but completely absent on US/UK keyboards. Space is a good alternative
that's easy to press on the vast majority of keyboards and doesn't have
a useful function in normal mode (it's equivalent to 'l'). Hopefully,
using space for both nvim and spacemacs doesn't mess up my muscle memory
too much.