Use xss-lock instead of systemd to trigger i3lock
This is the way that's recommended in the i3lock 2.14 changelog,
https://github.com/i3/i3lock/blob/01713948829dc36944cb79db89881b8ad56d5a37/CHANGELOG
(as well as in the man page, the GitHub readme, etc. for version 2.14),
because it ensures that the screen is locked *before* the computer
suspends / hibernates / etc. Additionally, unlike the systemd service I
was previously using, xss-lock also locks the screen when the monitor is
turned off due to user inactivity.