pacdiff: Use 1.5.0 Merge mode instead of pacdiff-merge3 hack
pacdiff 1.5.0 introduced a new Merge mode/action (see
https://gitlab.archlinux.org/pacman/pacman-contrib/-/blob/35e1e7f4730cd9b791fbbc88b3f57b83f911d0cb/CHANGES.md#150-2022-05-03).
This mode does most of the work of the pacdiff-merge3 script that I was
previously using, and it has some other advantages compared to
pacdiff-merge3, such as the ability to still view "regular" diffs
instead of merging - because the View action is no longer (ab)used to
perform a merge - and the addition of a step in which the merge result
can be reviewed before accepting or discarding it.

Even though Merge mode does most of the hard work (mainly, it determines
the previous package version and extracts of the old version of the file
from it), a pacdiff wrapper script is still required to make it work
nicely with kdiff3 and with AUR packages. Also, a wrapper script remains
a good place to specify default arguments for pacdiff, such as --backup
(see the previous commit) and --sudo (another 1.5.0 feature, which
obviates the need to run pacdiff itself with sudo).

The main downside of pacdiff's Merge mode is that it currently only
supports looking for old versions of packages in a single cache
directory per pacdiff command invocation. To make this work
simultaneously for both official and AUR packages, which (at least in my
case) are cached in different directories, my wrapper script creates a
single temporary directory containing symlinks to all packages.
Theoretically, this is not very efficient, because it creates many
symlinks (currently a few thousands) even though pacdiff usually only
needs to operate on a small number of packages. In practice, however,
this just takes a few milliseconds on my machine, so it should be fast
enough for any realistic number of installed packages.