xmobar: Replace custom volume script with Alsa monitor
Originally, I created the volume script as a non-polling alternative to
the polling 'Volume' monitor that xmobar provided at the time. Now, I've
just found out that xmobar 0.29 introduced the non-polling 'Alsa' volume
monitor (see
https://codeberg.org/xmobar/xmobar/src/commit/bfaa31d08229dc2c3dd1cef6064a322304569746/changelog.md#version-0-29-december-2018).
Hopefully this works well enough, otherwise this commit can be reverted
to restore the volume script (and maybe said script could be modified to
use an X property instead of a file, which might make it simpler and
more reliable).