xmonad: Choose profile-specific settings in Haskell code
Extensive use of dotdrop's templating feature weakens one of the main
strengths of Haskell: static type checking. If a different version of
the code is compiled for each profile, the only way to know that all of
the code compiles is to compile it for every profile.

This commit solves the above issue by only using templating to set a
single 'profile' variable, and then branching on this variable within
the Haskell code. This way, other than a trivial variable declaration,
the code that's compiled is the same for all profiles.

Some parts of the code are now slightly more complex (especially the
layout configuration), but the typical Haskell experience of "fearless"
editing (that this change will hopefully restore) should more than
compensate for this downside.