xmonad: Enable and fix -Wall warnings
-Wall is not the most complete set of warnings, but it's much better
than the default and, for a "simple" config file, it's also arguably
better than cluttering the top of the file with a huge pragma that
enables the ideal set of warnings.

Note that warnings aren't shown on the terminal when xmonad is
recompiled (unless there are deprecation warnings): instead, they're
redirected to $XMONAD_DATA_DIR/xmonad.errors.