xmonad: Create script which recompiles and shows warnings
As mentioned in commit acf1a5d060a89f30270a5907a568da804097f30a, xmonad
recompilation redirects warnings to a file, and then only shows the
contents of this file if there are errors and/or deprecation warnings,
but not if there are only other types of warnings. This script shows
said file whenever recompilation succeeds (i.e. there are no errors). An
unwanted consequence is that warnings will be shown twice when there are
deprecations, but it's probably not worth complicating the script to
avoid this.