fish: Switch from custom open function to default one
When I originally wrote my custom open function, the default fish open
function had the issue that it didn't background and disown xdg-open,
thus blocking the terminal, but this was then fixed in fish 3.2.0 (see
https://github.com/fish-shell/fish-shell/commit/5b706faa731ee073968dfa1d7b1f73375438c608).
Switching to the default function means that I can immediately enjoy any
improvements that to it that new fish versions may introduce, without
first having to manually merge them into my version of the function.

One difference between my open function and the (current version of the)
default one is that the former suppresses xdg-open's stdout and stderr,
while the latter doesn't. When everything goes well, this output is a
bit annoying, especially because it ends up on the same line as the
following shell prompt (due to xdg-open running in the background), but
when no program is available to open the specified file(s), then an
error message is better than a silent failure, so both suppressing and
not suppressing the output have upsides and downsides.