dotdrop: Create wrapper script which runs dotdrop in its virtualenv
This script not only circumvents the need to manually activate the
virtualenv before starting to use dotdrop (and to deactivate it when
done), but it also makes dotdrop usable from any directory even though
it's installed as a submodule of this repo.

Initially, I tried to create this as a fish function rather than a sh
script, because the dotdrop documentation recommends defining an alias
to make dotdrop usable from anywhere (see
https://dotdrop.readthedocs.io/en/latest/installation/#as-a-submodule),
and fish uses functions instead of aliases. However, since I also wanted
to locally source the virtualenv activation script, without affecting
the calling shell session, and since fish doesn't support subshells, I
had to pass the `source` and `dotdrop.sh` commands to a new fish
instance, in the form of properly escaped strings, and I couldn't figure
out how to correctly apply the `string escape` fish command to the
`$argv` variable (specifically, I ran into trouble with `string escape`
trying to interpret strings that start with '--' as options, and looking
at the man page, there doesn't seem to be an easy way around this).

Performance-wise, this script seems to have minimal overhead compared to
activating the virtualenv once and then running `dotdrop.sh` directly
(after the first activation, which seems to be slower than the following
ones, probably due to something like cache warmup effects).