.xinitrc: Move keyboard configuration to separate script
Lately, whenever some system packages update, the keyboard layout is
reset to the default until X is restarted (e.g. by logging out and back
in): https://bugs.archlinux.org/task/77789.

I've tried setting the keyboard layout in the X configuration files
(through localectl), but this didn't solve the issue, and instead caused
the keymappings I set with xmodmap in .xinitrc to stop working -- is
.xinitrc loaded before the X config files?

As a simple workaround, I'm moving the keyboard configuration to a
script that I can conveniently rerun to fix the layout settings whenever
an update messes them up.