dotdrop: Replace " with ' in config
Both single quotes (which delimit a "raw" string literal) and double
quotes (which allow escape sequences) would work in this case, but
`dotdrop import` replaces double quotes with single quotes when it edits
the config to add a new dotfile, thus introducing a change that should
either be reverted or committed. Instead of reverting it every time,
let's commit it once and for all.