.profile: Extract environment variable settings to separate file
On my current work machine, there already is a .profile file. I can edit
it, but I don't want to completely overwrite it, and I also don't want
to add the whole file to this dotfiles repo. Moving the part of .profile
that I actually need, i.e. environment variables, to a separate file
lets me install the latter file on the work machine and just add a
single "source" command to the (unversioned) .profile file.