fish: Provide $XDG_RUNTIME_DIR to bash when importing env
According to
https://manpages.ubuntu.com/manpages/xenial/en/man8/pam_systemd.8.html,
$XDG_RUNTIME_DIR is set by pam_systemd, rather than being defined in one
of the bash profile files. Therefore, it should already be present in
the environment when bash starts, otherwise any variable definitions (or
other things) that use $XDG_RUNTIME_DIR won't work correctly. On my work
machine, this resulted in the $DOCKER_HOST variable having the wrong
value.

Note that, if $XDG_RUNTIME_DIR wasn't defined on the system, fish would
automatically collapse the `XDG_RUNTIME_DIR=$XDG_RUNTIME_DIR` expression
to 0 arguments (due to the "cartesian product" behavior of variable
expansion), which is good, because it avoids introducing an undesirable
`XDG_RUNTIME_DIR=` definition into the environment.