spacemacs: Replace deprecated function in dotfile
Like all functions from the old CL library, `copy-list' is deprecated, and
should be replaced with `cl-copy-list'.

In theory, I would imagine that the deprecated function should still work.
However, for some reason, if Spacemacs is started in daemon mode, `copy-list' is
not yet defined at the time when the .spacemacs.d/init.el dotfile is loaded.
This results in an error that stops the remaining parts of the dotfile from
being loaded. Using the non-deprecated function fixes the issue.