configure-keyboard: Update for xkeyboard-config 2.43
As explained in
https://gitlab.freedesktop.org/xkeyboard-config/xkeyboard-config/-/merge_requests/694,
xkeyboard-config 2.43 maps the "Hyper" key to the "mod3" modifier by
default, whereas it was previously mapped to "mod4", i.e. the same
modifier that's produced by the Windows key (a.k.a. "Super") and used in
my xmonad config.

Due to the above change, this script no longer made caps lock act as an
additional Windows key when held. To fix it, I added an xmodmap command
that maps the caps lock key to the "mod4" modifier. (Another solution
would have been to replace every occurrence of "Hyper_L" with "Super_L"
in the script, but that would also have had the unintended side effect
of making the Windows key act like caps lock when pressed on its own.)

I also took the time to try and get a better understanding of how
xmodmap and xcape work and to add step-by-step explanations to the
commands I use in the script, since I've always found them quite
unintuitive. In fact, I still don't understand exactly how keys are
associated to modifiers: xmodmap's man page and modifier map output seem
to suggest that keycodes are directly bound to modifiers (although
keysyms are used as a convenient way to refer to keycodes without
hardcoding them), but in practice it seems that keysyms can somehow be
bound to modifiers too, otherwise the aforementioned xkeyboard-config
update wouldn't have affected this script (and I think the previous
version of the script wouldn't even have worked, since it remapped the
caps lock keycode to the "Hyper_L" keysym but didn't explicitly remap it
to the "mod4" modifier). Perhaps this is related to the differences
between xmodmap and XKB that the Arch wiki
(https://wiki.archlinux.org/title/Xmodmap) alludes to.

Finally, as part of improving readability, I moved the xmodmap command
that unbinds the "lock" modifier to the start of the script, so that I
can use the original "Caps_Lock" keysym instead of the remapped
"Hyper_L" to refer to the physical caps lock key's keycode, thus
reducing the layers of indirection that must be kept in mind in order to
understand what said command actually does.